Unified save/load functionality
