+++ primitives/__init__.py
@@ -1,2 +1,3 @@
 from primitives.signals import Signal, Side
+from primitives.StoredObject import StoredObject
 from primitives.trends import MarketTrend, TrendMovement
